% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_wills}
\alias{aggregate_wills}
\title{aggregate individual wills into a time series}
\usage{
aggregate_wills(
  wills_individual,
  aggregation = "1 weeks",
  aggname = aggregation,
  include_all = FALSE,
  start_weekday = NA
)
}
\arguments{
\item{wills_individual}{line-listed file containing individual wills}

\item{aggregation}{aggregation level}

\item{aggname}{text to display when referring to the aggregation level}

\item{start_weekday}{numeric value of weekday (1-7, Monday==1, but note that weekday assignment is wrong (but consistent) pre-Gregorian calendar anyway) on which to start binning (default NA means start on first day of dataset, whatever weekday that is)}
}
\value{
A \code{\link{data.frame}} that is also of class \code{epidemic.data}
}
\description{
Aggregate line-listed wills into a time series
of a specified temporal resolution.
}
\examples{
husting_wills_3weeks <- aggregate_wills(husting_wills_individual,"3 weeks","tri-weekly")
plotPlague(data=husting_wills_3weeks)

}
