% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregation.level}
\alias{aggregation.level}
\title{guess aggregation level of incidence data}
\usage{
aggregation.level(data)
}
\arguments{
\item{data}{data frame containing a \code{day} column}
}
\value{
a text label
}
\description{
guess aggregation level of incidence data
}
