% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggsum_monthly}
\alias{aggsum_monthly}
\title{Take data that are already in count form and change time unit
Should be faster than the other alternative}
\usage{
aggsum_monthly(year, month, day, deaths)
}
\description{
Take data that are already in count form and change time unit
Should be faster than the other alternative
}
