% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_draw}
\alias{ar1_draw}
\title{Randomly guess starting parameters for AR1}
\usage{
ar1_draw(draw_id, last_phi, sigma2_total, model_type)
}
\arguments{
\item{draw_id}{An \code{unsigned int} that contains the draw principles.}

\item{last_phi}{A \code{double} containing the last guessed phi value.}

\item{sigma2_total}{A \code{double} that contains the sum of all WVs.}

\item{model_type}{A \code{string} that describes the model transformation.}
}
\value{
A \code{vec} containing smart parameter starting guesses to be iterated over.
}
\description{
Sets starting parameters for each of the given parameters.
}
\examples{
#TBA
}
\keyword{internal}
