% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_draws}
\alias{arma_draws}
\title{Randomly guess starting parameters for ARMA}
\usage{
arma_draws(p, q, sigma2_total)
}
\arguments{
\item{p}{An \code{unsigned int} that contains the amount of AR parameters to generate.}

\item{q}{An \code{unsigned int} that contains the amount of MA parameters to generate.}

\item{sigma2_total}{A \code{double} that contains the sum of all WVs.}
}
\value{
A \code{vec} containing smart parameter starting guesses to be iterated over.
}
\description{
Sets starting parameters for each of the given parameters.
}
\examples{
#TBA
}
\keyword{internal}
