% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/inference.cpp, src/inference.h
\name{bootstrap_gof_test}
\alias{bootstrap_gof_test}
\title{Compute the Bootstrapped GoF Test}
\usage{
bootstrap_gof_test(obj_value, bs_obj_values, alpha, bs_gof_p_ci)
}
\arguments{
\item{obj_value}{A \code{double} that contains the optimized objective function value.}

\item{bs_obj_values}{A \code{vec} that contains the objective function values under bootstrap.}

\item{alpha}{A \code{double} that indicates the confidence.}

\item{bs_gof_p_ci}{A \code{bool} that indicates whether CIs should be included or not.}
}
\value{
A \code{vec} that has
\itemize{
\item Test Statistic
\item Low CI
\item Upper CI - BS
}
}
\description{
Handles the bootstrap computation and the bootstrapped p-value.
}
\keyword{internal}
