% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{brick_wall}
\alias{brick_wall}
\title{Removal of Boundary Wavelet Coefficients}
\usage{
brick_wall(x, wave_filter, method)
}
\arguments{
\item{x}{A \code{field<vec>} that contains the nlevel decomposition using either modwt or dwt.}

\item{wave_filter}{A \code{field<vec>} containing filter information. Only "haar" is implemented.}

\item{method}{A \code{string} to describe the mode. Choose between "modwt" and "dwt"}
}
\value{
A \code{field<vec>} with boundary modwt or dwt taken care of.
}
\description{
Removes the first n wavelet coefficients.
}
\details{
The vectors are truncated by removing the first n wavelet coefficients. 
These vectors are then stored into the field that is returned.
Note: As a result, there are no NA's introduced and hence the na.omit is not needed.
}
\examples{
x = rnorm(100)
me = modwt_cpp(x, filter_name = "haar", nlevels = 4, boundary = "periodic", brickwall = FALSE)
brick_wall(me, select_filter("haar"), "modwt")
}
\keyword{internal}
