% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_command.R, R/build_sm_command.R
\name{build_sm_command}
\alias{build_sm_command}
\title{Build SqueezeMeta command}
\usage{
build_sm_command(
  project,
  sample_file,
  input_dir,
  mode = "sequential",
  threads = 8
)

build_sm_command(
  project,
  sample_file,
  input_dir,
  mode = "sequential",
  threads = 8
)
}
\description{
Build SqueezeMeta command

Build SqueezeMeta command
}
