% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ci_eta3}
\alias{ci_eta3}
\title{Generate eta3 confidence interval}
\usage{
ci_eta3(y, dims, alpha_ov_2)
}
\arguments{
\item{y}{A \code{vec} that computes the brickwalled modwt dot product of each wavelet coefficient divided by their length.}

\item{dims}{A \code{String} indicating the confidence interval being calculated.}

\item{alpha_ov_2}{A \code{double} that indicates the \eqn{\left(1-p\right)*\alpha}{(1-p)*alpha} confidence level}
}
\value{
A \code{matrix} with the structure:
\itemize{
 \item{Column 1}{Wavelet Variance}
 \item{Column 2}{Chi-squared Lower Bounds}
 \item{Column 3}{Chi-squared Upper Bounds}
}
}
\description{
Computes the eta3 CI
}
\examples{
x = rnorm(100)
# Uses the internal MODWT function not associated with an S3 class.
decomp = modwt_cpp(x, filter_name = "haar", nlevels = 4, boundary = "periodic", brickwall = TRUE)
y = wave_variance(decomp)
ci_wave_variance(decomp, y, type = "eta3", alpha_ov_2 = 0.025)
}
\keyword{internal}
