% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dilution_effect_strength.R
\name{compute_dilution_effect_strength}
\alias{compute_dilution_effect_strength}
\title{Compute effect of increasing number of reservoir species}
\usage{
compute_dilution_effect_strength(m, nmax, ...)
}
\arguments{
\item{m}{Number of vector species}

\item{nmax}{Maximum number of reservoir species}

\item{...}{Additional params passed to \code{\link[=run_step_SIR_multihosts]{run_step_SIR_multihosts()}}}
}
\description{
Compute effect of increasing number of reservoir species
}
\examples{
eff <- compute_dilution_effect_strength(
  m = 10, nmax = 100, N = 1e5,
  tau = 0.5, phi_intra = 1/2, phi_inter = 1/4,
  b = 0.2, d = 0.2,
  sigma = 0.5, alpha = 0.9
)
plot(eff, type = "l")

}
