% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_models}
\alias{count_models}
\title{Count Models}
\usage{
count_models(desc)
}
\arguments{
\item{desc}{A \code{vector<string>} that contains the model's components.}
}
\value{
A \code{map<string, int>} containing how frequent the model component appears.
}
\description{
Count the amount of models that exist.
}
\keyword{internal}
