% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_cr.R
\name{cr_list}
\alias{cr_list}
\title{Split populations' clonotypes to their respective CR-subgroups.}
\usage{
cr_list(clonotype_list, ...)
}
\arguments{
\item{clonotype_list}{a character vector list of the clonotypes found in each individual group}

\item{...}{}
}
\value{
list of distinct clonotype groups named by the corisponding CR-types: 'private', 'exclusive', 'inclusive'
}
\description{
Make a list of the clonotypes based on the CR-types (private, exclusive & inclusive) that correspond to each one.
}
\examples{

clonotype_list <- rand_group(3, 3, 3)

cr_seq(clonotype_list)

}
