% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_overlap.R
\name{cr_overlap}
\alias{cr_overlap}
\title{Title}
\usage{
cr_overlap(subpopulations, func = intersect_percentage)
}
\arguments{
\item{subpopulations}{list}

\item{func}{finction}
}
\value{
matrix
}
\description{
Title
}
\examples{

g1 <- rand_group(n_sample=7, seq_n = 100, seq_l=5)
g2 <- rand_group(n_sample=12, seq_n = 100, seq_l=5)

sub_g1_i <- list(1:3, 4:7)
sub_g2_i <- list(1:5, 6:12)

cr1 <- cr_list(g1, "aa", sub_g1_i)
cr2 <- cr_list(g2, "aa", sub_g2_i)

names_g <- c("Cancer", "Control")
populations <- setNames(list(cr1, cr2), names_g)

cr_overlap(populations)

}
