% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_ymd}
\alias{date_to_ymd}
\title{convert dates to a data frame of year, month, and day}
\usage{
date_to_ymd(datetxt)
}
\arguments{
\item{datetxt}{dates as character}
}
\description{
convert dates to a data frame of year, month, and day
}
