\name{ddrescueInstructions}
\alias{ddrescueInstructions}
\alias{ddrescueInstructions.data.frame}
\alias{ddrescueInstructions.ggplot}
\title{Create instructions for data/plot recovering}
\description{
 Currently works for data.frame and plots
}
\usage{
ddrescueInstructions(object, archiveRead, md5hash) 

ddrescueInstructions.ggplot(object, archiveRead, md5hash) 

ddrescueInstructions.data.frame(object, archiveRead, md5hash) 
}
\author{
 Przemyslaw Biecek
}
\examples{
}

