% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deriv_2nd_dr}
\alias{deriv_2nd_dr}
\title{Analytic second derivative matrix for drift process}
\usage{
deriv_2nd_dr(tau)
}
\arguments{
\item{tau}{A \code{vec} containing the scales e.g. \eqn{2^{\tau}}{2^tau}}
}
\value{
A \code{matrix} with the first column containing 
the second partial derivative with respect to \eqn{\omega}{omega}.
}
\description{
To ease a later calculation, we place the result into a matrix structure.
}
\examples{
deriv_2nd_dr(2^(1:5))
}
\author{
James Joseph Balamuta (JJB)
}
