% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intgr_vec}
\alias{intgr_vec}
\alias{diff_inv_values}
\alias{diff_inv}
\title{Discrete Intergral: Inverse Difference}
\usage{
intgr_vec(x, xi, lag)

diff_inv_values(x, lag, d, xi)

diff_inv(x, lag, d)
}
\arguments{
\item{x}{A \code{vec} containing the data}

\item{xi}{A \code{vec} with length \eqn{lag*d} that provides initial values for the integration.}

\item{lag}{An \code{unsigned int} indicating the lag between observations.}

\item{d}{An \code{unsigned int} which gives the number of "differences" to invert.}
}
\description{
Takes the inverse difference (e.g. goes from diff() result back to previous vector)
}
\keyword{internal}
