% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_polyroot_arma}
\alias{do_polyroot_arma}
\title{Root Finding C++}
\usage{
do_polyroot_arma(z)
}
\arguments{
\item{z}{A \code{cx_vec} (complex vector) that has 1 in the beginning (e.g. c(1,3i,-3i))}
}
\description{
Used to interface with Armadillo
}
\keyword{internal}
