% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doublingTime,numeric-method}
\alias{doublingTime,numeric-method}
\title{compute doubling time from exponential growth rate}
\usage{
\S4method{doublingTime}{numeric}(obj)
}
\arguments{
\item{obj}{growth rate in units 1/year}
}
\description{
The doubling time in days is \code{log(2)/obj*365},
where \code{obj} is the exponential growth rate in years.
}
