% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epidemic_defs}
\alias{epidemic_defs}
\title{Epidemic year definitions}
\format{
A data frame containing columns:
\describe{
\item{\code{outbreak.year}}{numeric outbreak year}
\item{\code{severity}}{"major" or "minor"}
\item{\code{start}}{starting time (julian date)}
\item{\code{end}}{ending time (julian date)}
}
}
\usage{
epidemic_defs
}
\description{
Epidemic year definitions
}
\keyword{datasets}
