% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{class}
\name{epigrowthfit}
\alias{epigrowthfit}
\alias{getInits}
\title{epidemic growth fit}
\usage{
getInits(
  time,
  deaths,
  theta0 = NULL,
  peak = NULL,
  first = NULL,
  first_level = NULL,
  min_window_len = 5,
  distrib,
  model,
  peak_method = "max",
  skip_zero = TRUE,
  min_r = 0.1,
  debug_window = FALSE,
  debug_plot = FALSE
)

\S4method{initialize}{epigrowthfit}(
  .Object,
  time = NULL,
  deaths = NULL,
  data = NULL,
  time_var = "time",
  deaths_var = "plague.deaths",
  baseline = NULL,
  fixed = NULL,
  theta0 = NULL,
  first_time = NULL,
  first = NULL,
  first_level = 0.02,
  min_window_len = 5,
  peak = NULL,
  peak_extend = NULL,
  last = NULL,
  last_time = NULL,
  skip_zero = TRUE,
  link_vals = NULL,
  model = getOption("epigrowthfit.model", "richards"),
  distrib = getOption("epigrowthfit.distrib", "nbinom"),
  optimizer = getOption("epigrowthfit.optimizer", "user"),
  optimfun = getOption("epigrowthfit.optimfun", NULL),
  method = getOption("epigrowthfit.method", "L-BFGS-B"),
  upper = Inf,
  optCtrl = getOption("epigrowthfit.optCtrl", list()),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{time}{time vector}

\item{deaths}{death vector}

\item{theta0}{initial parameter guess}

\item{peak}{estimated peak of epidemic (fitting window ends at peak
for exponential models, peak+1 otherwise; initial parameter guesses
are based on log-linear fits to the data from the first to
halfway between the first step #' and the peak)}

\item{first}{first time step to use}

\item{first_level}{threshold to pick start of the fitting window (as a fraction of the peak incidence)}

\item{distrib}{observation error distribution to use (FIXME:
slot name should match argument?  (distrib != loglik)}

\item{model}{model to use}

\item{data}{data frame}

\item{time_var}{column name of time vector within \code{data}}

\item{deaths_var}{column name of deaths vector within \code{data}}

\item{baseline}{A "baseline" refers to a cumulative incidence model that incorporates
a \code{b*t} term.  If \code{baseline=TRUE} then (if a baseline is being used) 
its value (\code{b}) is fixed (i.e., it is a fixed offset), based on
the computed mean values in the surrounding years, rather than estimated
from the outbreak year data.  Whether or not a baseline is used is specified
by the \code{add_const} argument of the \code{\link{get_model}} function.}

\item{first_time}{start time for fitting window}

\item{last_time}{end time for fitting window}

\item{transforms}{a list of formulas specifying the parameter transformations, both for the model and for the log-likelihood function. Each transform must be a formula specifying a parameter depending on a smooth function of itself, e.g., x0~exp(x0).}

\item{inverses}{a list of formulas specifying the inverse parameter transformations, both for the model and for the log-likelihood function. Each transform must be a formula specifying a parameter depending on a smooth function of itself, e.g., x0~exp(x0).}
}
\value{
If \code{return_fit} is \code{FALSE}, a list containing elements \code{growth.rate} (point estimate (\code{value} and lower/upper CIs for the growth rate (\code{lower}, \code{upper})) and \code{fit} (predictions: \code{time}, observed \code{data}, and fitted values \code{fit}). If \code{return_fit} is \code{TRUE}, the results of \code{\link{exp_fit}}
}
\description{
A class representing an epidemic growth fit obtained
using the methodology of Ma \emph{et al.} (2014)

Fitting function for plague data. Uses either Richards (by default, or if an alpha parameter is supplied in \code{theta0}) or logistic model (if alpha not supplied, or if data set contains <= 4 points).

\code{getInits} picks an appropriate fitting window (see Details).
}
\details{
\describe{
\item{fitting window}{if \code{peak} is \code{NULL} and \code{first_level} is non-\code{NULL} (the default), \code{peak} is set to the maximum of the data, while \code{first} is set to the first element after the last observation with cases/deaths less than (\code{first_level*max(data)}). Otherwise, if \code{first_level} and \code{first} are both \code{NULL}, \code{first} is set to the first element after the last local minimum before the peak. If either \code{first} or \code{peak} are non-\code{NULL} (integers), they are set to the specified values. The fitting window goes from elements \code{first} to \code{peak}+1 (inclusive) of the data (or to \code{peak} for exponential fits).}
\item{initial parameter values}{If initial parameters (\code{theta0}) are not supplied, uses log-linear fitting over a range starting at \code{first} and going halfway to \code{peak+1} to estimate the initial size and growth rates. The initial population size (K) is set to the total number of cases/deaths. The Richards parameter (s) is initially set to 1.0001; the log of the negative binomial size parameter is initially set to 1.}
\item{data frame structure}{Two \code{numeric} columns are required, which by default are assumed to be named \code{"time"} and \code{"plague.deaths"}.  Either column name can be changed using the \code{time_var} and \code{deaths_var} arguments.}
\item{debugging}{to debug, use \code{trace("initialize",sig="epigrowthfit",browser)}}
}
}
\section{Slots}{

\describe{
\item{\code{model}}{the trajectory model (an object of class \code{\link{Model}})}

\item{\code{loglik}}{the log-likelihood for the observation error distribution}

\item{\code{time}}{decimal dates at each point in the time series}

\item{\code{date}}{time as \code{\link{Date}} object}

\item{\code{deaths}}{deaths in the time series}

\item{\code{window}}{the fitting window (indices of times)}

\item{\code{mle2}}{an \code{\link{mle2}} object}

\item{\code{growthRate}}{the fitted exponential growth rate}

\item{\code{mean}}{predicted mean value?}
}}

\examples{
## extract Great Plague of London from full plague data frame
great.plague <- subset(london_bills,outbreak.year==1665)
## run default fit, which uses Richards model
gpfit.richards <- epigrowthfit(data=great.plague)
## run fit using logistic model
gpfit.logistic <- epigrowthfit(data=great.plague,model="logistic")
## run fit using exponential model
gpfit.exp <- with(great.plague,epigrowthfit(time,plague.deaths,model="exp"))
## print growth rates and confidence intervals for each fit
growthRate(gpfit.richards)
growthRate(gpfit.logistic)
growthRate(gpfit.exp)
## print associated doubling time, R0 and fitting window range for Richards model fit
doublingTime(gpfit.richards)
R0(gpfit.richards)
range(window(gpfit.richards))
gpfit.richards
## spew full summary of this fit
summary(gpfit.richards)
## plot fit in various ways
plot(gpfit.richards)
plot(gpfit.richards,log="y")
plot(gpfit.richards,log="y",window.only=TRUE)
plot(gpfit.richards,cumulative=TRUE)
plot(gpfit.richards,cumulative=TRUE,log="y")
## plot all fits on same graph for comparison
plot(gpfit.richards,lwd.pred=6,show.legend=FALSE,show.gof=FALSE)
plot(gpfit.logistic,add=TRUE,col.pred="green",lwd.pred=1,show.data=FALSE)
plot(gpfit.exp,add=TRUE,col.pred="blue",show.data=FALSE)
## add legend showing goodness of fit of each model
legend.text <- sprintf("\%s (\%.3f)",
  c("Richards","logistic","exp"),
  c(gof(gpfit.richards),gof(gpfit.logistic),gof(gpfit.exp)))
legend("topright",bty="n",legend=legend.text,lty=1,lwd=c(6,1,3),col=c("black","green","blue"))
title(main="Initial Growth Fits to Great Plague of London in 1665")
\dontrun{
    ## hard-code fitting window
    gpfit.3 <- update(gpfit.logistic, first_time=1665.5, last_time=1666.0)
    ## constant-baseline fits
    d1375 <- subset(husting_wills,outbreak.year==1375)
    gpfit.4 <- epigrowthfit(data=d1375,
          model=get_model("logistic",add_const=TRUE),
          first=1,peak_extend=4,
          deaths_var="nwills",
          baseline=TRUE)
}
if (require("outbreaks")) {
#' ## example using outbreaks package data
}
}
\references{
\insertRef{Ma+14}{epigrowthfit}
}
