% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fit.R
\name{exp_fit}
\alias{exp_fit}
\title{exponential fit}
\usage{
exp_fit(
  t,
  X,
  theta0,
  model,
  loglik = get_loglik("nbinom"),
  confint = TRUE,
  fixed = NULL,
  lower = NULL,
  upper = NULL,
  optimizer = "user",
  optimfun = NULL,
  method = "L-BFGS-B",
  optCtrl = list(),
  extra_optim = FALSE,
  fallback_stderr_fac = 0.1,
  drop_mle2_call = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{t}{time vector}

\item{X}{vector of cases/deaths/etc.}

\item{theta0}{A named vector of initial parameter values}

\item{model}{the model to fit to}

\item{loglik}{the log-likelihood function}

\item{confint}{(logical) generate confidence interval?}

\item{lower}{lower bounds (see \code{link{bbmle::mle2}})}

\item{upper}{upper bounds (see \code{link{bbmle::mle2}})}

\item{optimizer}{name of optimization function, e.g. \sQuote{optim} or \sQuote{nlminb} (see \code{\link{mle2}})}

\item{optimfun}{user-specified optimization function (see \code{\link{mle2}})}

\item{method}{name of optimization method (ignored unless \code{optimizer} is "optim")}

\item{fallback_stderr_fac}{quasi-standard error (for profiling), as a proporition of estimated r value, when Hessian is non-positive-definite}

\item{drop_mle2_call}{to save space, drop call slot from mle2 object}
}
\description{
exponential fit
}
\examples{
dd <- subset(london_bills,outbreak.year==1665)
ff <- exp_fit(t=dd$time[15:40], X=dd$plague.deaths[15:40],
        theta0=c(r=3,x0=-10,K=10,ll.k=0),
        model=get_model("logistic"),
        loglik=get_loglik("nbinom"))
ffb <- exp_fit(t=dd$time[15:40], X=dd$plague.deaths[15:40],
        theta0=c(r=3,x0=-10,K=10,ll.k=0,b=0),
        model=get_model("logistic",add_const=TRUE),
        loglik=get_loglik("nbinom"),
        optimizer="nlminb",
        fixed = list(b=-1))
}
