% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_full_model}
\alias{find_full_model}
\title{Find the Common Denominator of the Models}
\usage{
find_full_model(x)
}
\arguments{
\item{x}{A \code{vector< vector<string> >} that contains all possible models under consideration}
}
\value{
A \code{vector<string>} that contains the terms of the common denominator of all models
}
\description{
Determines the common denominator among models
}
\keyword{internal}
