% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/inference.cpp, src/inference.h
\name{format_ci}
\alias{format_ci}
\title{Format the Confidence Interval for Estimates}
\usage{
format_ci(theta, se, alpha)
}
\arguments{
\item{theta}{A \code{vec} containing the estimates}

\item{se}{A \code{vec} containing the standard error}

\item{alpha}{A \code{double} that contains the confidence level.}
}
\value{
A \code{mat} that has:
\itemize{
\item Column 1: Lo CI
\item Column 2: Hi CI
\item Column 3: SE
}
}
\description{
Creates hi and lo confidence based on SE and alpha.
}
\keyword{internal}
