% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_model}
\alias{gen_model}
\title{Generate Time Series based on Model (Internal)}
\usage{
gen_model(N, theta, desc, objdesc)
}
\arguments{
\item{N}{An \code{interger} containing the amount of observations for the time series.}

\item{theta}{A \code{vec} containing the parameters to use to generate the model}

\item{desc}{A \code{vector<string>} containing the different model types (AR1, WN, etc..)}

\item{objdesc}{A \code{field<vec>} contains the different model objects e.g. AR1 = c(1,1)}
}
\value{
A \code{vec} that contains combined time series.
}
\description{
Create a time series process based on a supplied \code{ts.model}.
}
\examples{
# AR
set.seed(1336)
gen_model(1000, c(.9,1), "AR1", list(c(1,1)))
}
\keyword{internal}
