% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generation.interval.plague}
\alias{generation.interval.plague}
\title{Generation interval distribution for pneumonic plague}
\usage{
generation.interval.plague()
}
\description{
compute generation interval for plague by convolving
(built-in) latent period distribution with reversed
infectious period distribution
}
\references{
\insertRef{Sven07}{epigrowthfit}
}
