% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline.R
\name{get_baseline}
\alias{get_baseline}
\title{Get baseline}
\usage{
get_baseline(
  data,
  epid_defs,
  in_col = "total",
  out_col = "plague.deaths",
  time_col = "time",
  per_wid = 2,
  min_val = 0,
  include_all = FALSE,
  do_plot = FALSE,
  y_min = 0.1
)
}
\arguments{
\item{data}{data frame including at least columns \code{time} and \code{in_col}}

\item{epid_defs}{data frame including outbreak.year, start, end}

\item{in_col}{name of deaths column in input}

\item{out_col}{name of deaths column in output}

\item{time_col}{name of numeric time column in input}

\item{per_wid}{number of years on either side of epidemic to include baseline}

\item{min_val}{minimum value (floor) for baseline}

\item{include_all}{include all values, not just epidemic-year data (i.e. time series of counts between epidemics) in the returned data frame?}

\item{do_plot}{produce diagnostic plots?}

\item{y_min}{minimum y-value for plotting}
}
\value{
a data frame of similar structure to the one that was input as \code{data}, with the following modifications:
\itemize{
\item new columns \code{prev<i>} and \code{next<i>} (where <i> runs from 1 to \code{perwid} representing the counts in previous and succeeding years; these are useful for diagnostic plots
\item a column labelled \code{out_col} (the difference between \code{in_col}
and the computed baseline)
\item an attribute \code{baseline}: a named numeric vector with the baseline value for each epidemic in \code{epid_defs} (names correspond to the \code{outbreak.year} variable in \code{epid_defs}
}
}
\description{
add an estimated deaths column that is the total deaths minus a baseline estimated by subtracting the average mortality in surrounding years for exactly the part of the year we are using for our fits
}
\details{
for the method to work properly, the data should include values for the periods before and after each outbreak year included in \code{epidemic_defs} (depends on \code{per_wid}: +/- 2 years by default)
}
\examples{
\dontrun{
library(dplyr)
ctmp <- canterbury_wills_individual \%>\%
   get_date_count("2 weeks") \%>\%
   get_baseline(epidemic_defs,in_col="nwills",
                do_plot=FALSE,include_all=TRUE)
}
}
