% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date_count}
\alias{get_date_count}
\title{tabulate numbers of wills & testaments}
\usage{
get_date_count(
  data,
  unit = "1 days",
  date_col = "date",
  count_col = "nwills",
  start_weekday = NA,
  start_date = NA
)
}
\arguments{
\item{unit}{time unit for aggregation}

\item{date_col}{name of date column}

\item{count_col}{name of aggregated column}

\item{start_weekday}{starting weekday for aggregation}

\item{start_date}{starting date for aggregation}

\item{dd}{data frame}
}
\description{
tabulate numbers of wills & testaments
}
