% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_infected.R
\name{get_max_infected}
\alias{get_max_infected}
\title{Get total number of infected in an epidemic}
\usage{
get_max_infected(m, n, N, phi_intra, phi_inter, ...)
}
\arguments{
\item{m}{Number of vector species}

\item{n}{Number of reservoir species}

\item{...}{Additional params passed to \code{\link[=run_step_SIR_multihosts]{run_step_SIR_multihosts()}}}
}
\description{
Get total number of infected in an epidemic
}
\examples{
get_max_infected(
  m = 10, n = 10, N = 10e5,
  tau = 0.5, phi_intra = 1/2, phi_inter = 1/4,
  b = 0.5, d = 0.5,
  sigma = 0.5, alpha = 0.5
)

}
