% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline.R
\name{get_range}
\alias{get_range}
\title{Get range}
\usage{
get_range(ep, data, maxlen = Inf, time_col = "time")
}
\arguments{
\item{ep}{endpoints (in decimal time)}

\item{data}{data frame}

\item{maxlen}{max window length (no longer used)}

\item{time_col}{name of time variable}
}
\description{
Get range
}
