\name{getBitLy}
\alias{getBitLy}
\title{Use bit.ly API to shorten the link}
\description{
 Read more about BitLy API here: \link{\code{http://dev.bitly.com/}}
 
 You need to obtain generic token in order to use this function.
 Read more about 'OAuth Basic Authentication Flow' here: \link{\code{http://dev.bitly.com/authentication.html}}
}
\usage{getBitLy(url, mytoken)}
\author{
 Przemyslaw Biecek
}
\examples{
# getBitLy("http://smarterpoland.pl/", mytoken)
}
\keyword{datasets}
