% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR.R
\docType{methods}
\name{getExperimentData}
\alias{getExperimentData}
\alias{getExperimentData,RNAmodR-method}
\title{getExperimentData}
\usage{
getExperimentData(x, number = FALSE)

\S4method{getExperimentData}{RNAmodR}(x, number = FALSE)
}
\arguments{
\item{x}{a RNAmodR object}

\item{number}{optional number to return results for one experiment only}
}
\value{
a list containg the information on the experiment or a data.frame 
with all the information
}
\description{
returns setup data for the experiments based on the input
layout csv-file.
}
\examples{
\donttest{
data("RNAmodR_test", package = "RNAmodR")

# All experiments
getSampleData(experiment)
# One experiment as list
getSampleData(experiment, 1)
# Two experiments as data.frame
getSampleData(experiment, c(1,2))
}
}
