% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR-plot.R
\docType{methods}
\name{getModPlot}
\alias{getModPlot}
\alias{saveModPlot}
\alias{getModPlot,RNAmodR,numeric,character-method}
\alias{getModPlot}
\alias{saveModPlots}
\alias{saveModPlotFromSummarizedExperiment}
\alias{getModPlot,SummarizedExperiment,missing,character-method}
\alias{saveModPlot,RNAmodR,numeric,character-method}
\alias{saveModPlot,SummarizedExperiment,missing,character-method}
\title{Visualization of modifications}
\usage{
getModPlot(x, number, modifications, gene, gff, fasta, focus = FALSE)

saveModPlot(x, number, modifications, genes, gff, fasta, focus = FALSE,
  folder, filetype = "pdf")

\S4method{getModPlot}{RNAmodR,numeric,character}(x, number, modifications, gene,
  focus)

\S4method{getModPlot}{SummarizedExperiment,missing,character}(x, modifications,
  gene, gff, fasta, focus)

\S4method{saveModPlot}{RNAmodR,numeric,character}(x, number, modifications,
  genes, focus, filetype)

\S4method{saveModPlot}{SummarizedExperiment,missing,character}(x, modifications,
  genes, gff, fasta, focus, folder, filetype)
}
\arguments{
\item{x}{a RNAmod object or a SummarizedExperiment containg the experimental data.}

\item{number}{an experiemnt number, if \code{x} is RNAmodR}

\item{modifications}{name of modification to be used for analysis.}

\item{gene}{a single gene name used by \code{getModPlot}}

\item{gff}{a GRanges object with the genomic annotation data. Mandatory, if 
\code{x} is RNAmodR}

\item{fasta}{a FaFile object with the genomic sequences, if 
\code{x} is RNAmodR}

\item{focus}{optional: logical, whether to plot only the region next to the 
modification}

\item{genes}{gene names used by \code{saveModPlots}}

\item{folder}{folder for saving the output used by \code{saveModPlots}}

\item{filetype}{file type used for \code{saveModPlots}}
}
\value{
\code{getModPlot}: a \code{TableGrob} of layer plot and any mod plot

\code{saveModPlot}: TRUE if successful
}
\description{
\code{getModPlot} and \code{saveModPlot} plot the results by combining
position data and modifications with transcript layout data.
}
\examples{
\donttest{
getModPlot(mod,1,"m7G","RDN18-1")
saveModPlot(mod,1,"m7G",c("RDN18-1","tS\\\\(CGA\\\\)P"))
}
}
