% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{gmwm-defunct}
\alias{gmwm-defunct}
\alias{compare.gmwm}
\alias{compare.models}
\alias{compare.wvar}
\alias{gen.gts}
\alias{gen.lts}
\alias{gmwm.imu}
\alias{auto.imu}
\alias{rank.models}
\alias{demo.lts}
\title{Defunct function(s) in the gmwm package}
\usage{
compare.gmwm(...)

compare.wvar(...)

compare.models(...)

gen.gts(...)

gen.lts(...)

demo.lts(...)

gmwm.imu(...)

rank.models(...)

auto.imu(...)
}
\arguments{
\item{...}{Grab old parameters}
}
\description{
These functions have been removed from the gmwm package.
}
\section{Details}{

\tabular{rl}{
  \code{compare.gmwm} \tab has been removed in favor of \code{\link{compare_models}}\cr
  \code{compare.models} \tab has been removed in favor of \code{\link{compare_models}}\cr
  \code{compare.wvar} \tab has been removed in favor of \code{\link{compare_wvar}}\cr
  \code{gen.gts} \tab has been removed in favor of \code{\link{gen_gts}}\cr
  \code{gen.lts} \tab has been removed in favor of \code{\link{gen_lts}}\cr
  \code{demo.lts} \tab has been removed in favor of \code{\link{demo_lts}}\cr
  \code{gmwm.imu} \tab has been removed in favor of \code{\link{gmwm_imu}}\cr
  \code{rank.models} \tab has been removed in favor of \code{\link{rank_models}}\cr
  \code{auto.imu} \tab has been removed in favor of \code{\link{auto_imu}}\cr
}
}

