% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/inference.cpp, src/inference.h
\name{gof_test}
\alias{gof_test}
\title{Compute the GOF Test}
\usage{
gof_test(theta, desc, objdesc, model_type, tau, v_hat, wv_empir)
}
\arguments{
\item{theta}{A \code{vec} with dimensions N x 1 that contains user-supplied initial values for parameters}

\item{desc}{A \code{vector<string>} indicating the models that should be considered.}

\item{objdesc}{A \code{field<vec>} containing a list of parameters (e.g. AR(1) = c(1,1), ARMA(p,q) = c(p,q,1))}

\item{model_type}{A \code{string} that contains the model type: \code{"imu"} or \code{"ssm"}}

\item{tau}{A \code{vec} containing the scales of a proccess.}

\item{v_hat}{A \code{mat} that contains the bootstrapped matrix.}

\item{wv_empir}{A \code{vec} that contains the empirical wavelet variance.}
}
\value{
A \code{vec} that has
\itemize{
\item Test Statistic
\item P-Value
\item DF
}
}
\description{
yaya
}
\keyword{internal}
