\name{graphID-class}
\docType{class}
\alias{graphID-class}
\title{Unique Graph Identifications}
\description{This class allows nodes and edges to have unique
  identifiers}
\author{Robert Gentleman}
\seealso{\code{\link[Ruuid]{Ruuid-class}} in package \code{Ruuid}.}

%\examples{
%##---- Should be DIRECTLY executable !! ----
%}
\keyword{classes}
