% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-ident-D.R, R/class-RNAmodR-ident-m3C.R,
%   R/class-RNAmodR-ident-m7G.R
\docType{methods}
\name{identifyModificationPerTranscript}
\alias{identifyModificationPerTranscript}
\alias{identifyModificationPerTranscript}
\alias{scoreModificationsPerTranscript,RNAmodRident_D,numeric,GRangesList,RNAmodRargs,numeric-method}
\alias{scoreModificationsPerTranscript,RNAmodRident_m3C,numeric,GRangesList,RNAmodRargs,numeric-method}
\alias{scoreModificationsPerTranscript,RNAmodRident_m7G,numeric,GRangesList,RNAmodRargs,numeric-method}
\title{Returns the data for each transcript}
\usage{
scoreModificationsPerTranscript(x, locations, data, args, endLocation)

identifyModificationsPerTranscript(x, data, args)


  \S4method{scoreModificationsPerTranscript}{RNAmodRident_D,numeric,GRangesList,RNAmodRargs,numeric}(x,
  locations, data, args, endLocation)


  \S4method{scoreModificationsPerTranscript}{RNAmodRident_m3C,numeric,GRangesList,RNAmodRargs,numeric}(x,
  locations, data, args, endLocation)


  \S4method{scoreModificationsPerTranscript}{RNAmodRident_m7G,numeric,GRangesList,RNAmodRargs,numeric}(x,
  locations, data, args, endLocation)
}
\description{
\code{RNAmodRident_D}

\code{RNAmodRident_m3C}

\code{RNAmodRident_m7G}
}
