% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{initialize,Model-method}
\alias{initialize,Model-method}
\title{the initializer for Model}
\usage{
\S4method{initialize}{Model}(
  .Object,
  name,
  model,
  input,
  par = NULL,
  inits = rep(NA_real_, length(par)),
  keep_hessian = FALSE
)
}
\arguments{
\item{model}{the formula specifying the model}

\item{input}{the input variable name}

\item{par}{the parameter names}

\item{inits}{initial values}

\item{keep_hessian}{maintain the hessian as part of the model?}
}
\description{
the initializer for Model
}
