% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{initialize,TransformedModel-method}
\alias{initialize,TransformedModel-method}
\title{initializer for TransformedModel}
\usage{
\S4method{initialize}{TransformedModel}(.Object, model, transforms = NULL, inverses = NULL)
}
\arguments{
\item{model}{the model to be transformed}

\item{transforms}{the parameter transforms, a list of formula}

\item{inverses}{the inverse transforms, a list of formula}
}
\description{
initializer for TransformedModel
}
