% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fit.R
\name{is.valid}
\alias{is.valid}
\title{check whether a variance-covariance matrix is valid}
\usage{
is.valid(V, tol = 1e-10)
}
\arguments{
\item{V}{the vcov matrix}

\item{tol}{tolerance for positive definiteness}
}
\value{
boolean
}
\description{
check whether a variance-covariance matrix is valid
}
