% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iswhole.R
\name{is.whole}
\alias{is.whole}
\title{Integer Check}
\usage{
is.whole(x)
}
\arguments{
\item{x}{A \code{numeric} value.}
}
\value{
A \code{boolean} value indicating whether the value is an integer or not.
}
\description{
Checks whether the submitted value is an integer
}
\examples{
is.whole(2.3)
is.whole(4)
is.whole(c(1,2,3))
is.whole(c(.4,.5,.6))
is.whole(c(7,.8,9))
}
\author{
James Balamuta
}
\keyword{internal}
