% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit2}
\alias{logit2}
\title{Logit2 Function}
\usage{
logit2(x)
}
\arguments{
\item{x}{A \code{vec} containing probabilities (e.g. -2 <= x <= 2)}
}
\value{
A \code{vec} containing logit terms.
}
\description{
This function computes the logit2 link function.
}
\examples{
# Set seed for reproducibility
set.seed(1142)

# Simulate data
x.sim = runif(10, -2, 2)

# Transform
x.sim.transformed = logit2(x.sim)
}
\seealso{
\code{\link{logit2_inv}}
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
