% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit_inv}
\alias{logit_inv}
\title{Logit Inverse Function}
\usage{
logit_inv(x)
}
\arguments{
\item{x}{A \code{vec} containing real numbers.}
}
\value{
A \code{vec} containing logit probabilities.
}
\description{
This function computes the inverse of a logit transformation of the parameters.
}
\examples{
# Set seed for reproducibility
set.seed(1412)

# Simulate data
x.sim = runif(10, -1, 1)

# Transform
x.sim.transformed = logit(x.sim)

# Untransform
x.sim.untransformed = logit_inv(x.sim.transformed)

# Compare results
results = cbind(x.sim, x.sim.untransformed)
}
\seealso{
\code{\link{logit}}
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
