% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mod_squared_cpp}
\alias{Mod_squared_cpp}
\title{Absolute Value or Modulus of a Complex Number Squared.}
\usage{
Mod_squared_cpp(x)
}
\arguments{
\item{x}{A \code{cx_vec}.}
}
\value{
A \code{vec} containing the modulus squared for each element.
}
\description{
Computes the squared value of the Modulus.
}
\details{
Consider a complex number defined as: \eqn{z = x + i y} with real \eqn{x} and \eqn{y},
The modulus is defined as: \eqn{r = Mod\left(z\right) = \sqrt{\left(x^2 + y^2\right)}}{r = Mod(z) = sqrt(x^2 + y^2)}
This function will return: \eqn{r^2 = Mod\left(z\right)^2 = x^2 + y^2}
}
\examples{
Mod_squared_cpp(c(1+.5i, 2+1i, 5+9i))
}
\keyword{internal}
