% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\alias{Model}
\title{a class representing a trajectory model used to fit exponential growth rates}
\description{
a class representing a trajectory model used to fit exponential growth rates
}
\section{Slots}{

\describe{
\item{\code{expr}}{an expression specifying the model}

\item{\code{grad}}{gradient with respect to the parameters}

\item{\code{hessian}}{Hessian of the model}

\item{\code{inits}}{initial values}

\item{\code{input}}{input variable name}

\item{\code{name}}{short name of the model}

\item{\code{output}}{output variable name}

\item{\code{par}}{parameter names}
}}

