\name{hasMultilevelModel-methods}
\docType{methods}
\alias{hasMultilevelModel-methods}
\alias{hasMultilevelModel,mrp-method}
\title{Indicates whether a multilevel model exists for the object}
\description{Returns true if the object has a multilevel model. This indicates 
    that the multilevel regression step has already been completed. Note: the 
    object does not do anything smart. This method does not indicate whether
    or not the multilevel model matches the current formula.}
\keyword{methods}