% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num_rep}
\alias{num_rep}
\title{Replicate a Vector of Elements \eqn{n} times}
\usage{
num_rep(x, n)
}
\arguments{
\item{x}{A \code{vec} containing the data}

\item{n}{An \code{unsigned int} indicating the number of times the vector should be repeated.}
}
\value{
A \code{vec} with repeated elements of the initial supplied vector.
}
\description{
This function takes a vector and replicates all of the data \eqn{n} times
}
\keyword{internal}
