% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fit.R
\name{optim_repeat_nlminb}
\alias{optim_repeat_nlminb}
\title{repeat nlminb until converged}
\usage{
optim_repeat_nlminb(
  par,
  fn,
  gr = NULL,
  ...,
  method = NULL,
  lower = -Inf,
  upper = Inf,
  control = list(),
  hessian = FALSE,
  strict = TRUE,
  max.tries = 25
)
}
\arguments{
\item{par}{parameter}

\item{fn}{function}

\item{gr}{gradient}

\item{method}{stub}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{control}{control arguments}

\item{hessian}{stub}

\item{strict}{?}

\item{max.tries}{max number of attempts}
}
\description{
repeat nlminb until converged
}
