% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{parseMod}
\alias{parseMod}
\title{parseMod}
\usage{
parseMod(object, gff, fafile, modClasses)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object for the genome}

\item{fafile}{a FaFile object for the genome}

\item{modClasses}{list of modification classes used for parsing}
}
