% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.r
\name{pkg_cmd}
\alias{pkg_cmd}
\alias{pkg_install}
\alias{pkg_build}
\alias{pkg_check}
\alias{pkg_name}
\title{Utility package functions}
\usage{
pkg_install(pkg, path = tempdir())

pkg_build(pkg, path = tempdir())

pkg_check(pkg, path = tempdir(), as_cran = FALSE)

pkg_name(pkg)
}
\arguments{
\item{pkg}{a path to package source (`pkg_install()`, `pkg_build()`, `pkg_name()`)
or tarball (`pkg_install()`, `pkg_check()`)}

\item{path}{a path where package is installed, build, or checked, defaults to `tempdir()`}

\item{as_cran}{whether to run `R CMD check` with the `--as-cran` preset}
}
\value{
`pkg_install` and `pkg_check` do not have a return value,
        `pkg_build` returns a path to the compiled package tarball,
         `pkg_name` returns a name of the package in specified path
}
\description{
These functions map to the `R CMD INSTALL`, `R CMD build` and `R CMD check` commands.
}
\details{
The functions `pkg_install`, `pkg_build`, and `pkg_check` map to the respective `R CMD`
commands. By default, all output files are created in a temporary directory.
The `pkg_name` is an utility functions that reads the `DESCRIPTION` file and returns the
package name.
}
