% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{print.dwt}
\alias{print.dwt}
\title{Print Discrete Wavelet Transform}
\usage{
\method{print}{dwt}(x, ...)
}
\arguments{
\item{x}{A \code{dwt} object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints the dwt decomposition
}
\description{
Prints the results of the modwt list
}
\examples{
set.seed(999)
x = rnorm(2^8)
dwt(x)
}
\author{
JJB
}
\keyword{internal}
