% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudo_logit}
\alias{pseudo_logit}
\title{Pseudo Logit Function}
\usage{
pseudo_logit(x)
}
\arguments{
\item{x}{A \code{vec} containing probabilities (e.g. 0 <= x <= 1)}
}
\value{
A \code{vec} containing logit terms.
}
\description{
This function compute the link function to constrain parameters to a positive domain.
}
\examples{
# Set seed for reproducibility
set.seed(1412)

# Simulate data
x.sim = runif(10, -1, 1)

# Transform
x.sim.transformed = pseudo_logit(x.sim)
}
\seealso{
\code{\link{pseudo_logit_inv}}
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
