% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudo_logit_inv}
\alias{pseudo_logit_inv}
\title{Pseudo Logit Inverse Function}
\usage{
pseudo_logit_inv(x)
}
\arguments{
\item{x}{A \code{vec} containing real numbers.}
}
\value{
A \code{vec} containing logit probabilities.
}
\description{
This function computes the pseudo inverse of a logit transformation of the parameters in order to constrain them to a positive domain
}
\examples{
# Set seed for reproducibility
set.seed(1222)

# Simulate data
x.sim = runif(10, -1, 1)

# Transform
x.sim.transformed = pseudo_logit(x.sim)

# Untransform
x.sim.untransformed = pseudo_logit_inv(x.sim.transformed)

# Compare results
results = cbind(x.sim, x.sim.untransformed)
}
\seealso{
\code{\link{pseudo_logit}}
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
