% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_cr.R
\name{public_cr}
\alias{public_cr}
\title{Label a clonotype list with the appropriate CR classifications.}
\usage{
public_cr(share_tbl)
}
\arguments{
\item{clonotype_list}{list of character vectors}
}
\value{
character vector named by colonotypes
}
\description{
clonotypes are given a label depending on the CR-types that they are associated with.
}
\examples{

rand_rep <- function(n, alphabet = LETTERS[1:10])
  sample(alphabet, n, replace = TRUE)

rep_list <- replicate(3, rand_rep(rpois(1, 10)), simplify = FALSE)

public_cr(rep_list)

}
