% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qmf}
\alias{qmf}
\title{Quadrature Mirror Filter}
\usage{
qmf(g, inverse)
}
\arguments{
\item{g}{A \code{vector} that contains the filter constants.}

\item{inverse}{A \code{bool} that indicates whether the inverse quadrature mirror filter is computed. 
By default, the inverse quadrature mirror is computed.}
}
\value{
A \code{vector} that contains either the forward QMF (evalute in order) or the inverse QMF (reverse order).
}
\description{
Calculate the series quadrature mirror filter (QMF). Requires a series of an even length.
}
\examples{
# Haar values
g = rep(1/sqrt(2),2)
qmf(g)
}
\author{
JJB
}
\keyword{internal}
