% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigrowthfit.R
\docType{methods}
\name{R0,epigrowthfit-method}
\alias{R0,epigrowthfit-method}
\title{returns R0 and its 95\% confidence interval}
\usage{
\S4method{R0}{epigrowthfit}(obj, generation.interval = generation.interval.plague())
}
\arguments{
\item{obj}{a fitted model object}

\item{generation.interval}{the generation interval distribution}
}
\value{
a vector
}
\description{
returns R0 and its 95\% confidence interval
}
\references{
\insertRef{WallLips07}{epigrowthfit}
}
