% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_ARIMA}
\alias{Rcpp_ARIMA}
\title{Hook into R's ARIMA function}
\usage{
Rcpp_ARIMA(data, params)
}
\arguments{
\item{data}{A \code{vec} of data.}

\item{params}{A \code{vec} of the ARMA parameters}
}
\value{
A \code{vec} containing the CSS of the ARMA parameters.
}
\description{
Uses R's ARIMA function to obtain CSS values for starting condition
}
\keyword{internal}
