% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmodR-ident.R,
%   R/class-RNAmodR-ident-D.R, R/class-RNAmodR-ident-m3C.R,
%   R/class-RNAmodR-ident-m7G.R, R/class-RNAmodR-ident-ribometh.R
\docType{class}
\name{RNAmodR-ident-class}
\alias{RNAmodR-ident-class}
\alias{show,RNAmodRident-method}
\alias{RNAmodRident_D-class}
\alias{RNAmodRident_m3C-class}
\alias{RNAmodRident_m7G-class}
\alias{RNAmodRident_RiboMeth-class}
\title{RNAmodR identification class}
\usage{
\S4method{show}{RNAmodRident}(object)
}
\arguments{
\item{object}{a RNAmodRident class object}
}
\description{
Virtual class for modification detection.

\code{RNAmodRident_D}:

\code{RNAmodRident_m3C}:

\code{RNAmodRident_m7G}:

\code{RNAmodRident_RiboMeth}:
}
\section{Slots}{

\describe{
\item{\code{analysisType}}{}

\item{\code{modType}}{}
}}

