% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtruncated_normal}
\alias{rtruncated_normal}
\title{Truncated Normal Distribution Sampling Algorithm}
\usage{
rtruncated_normal(n, mu, sigma, a, b)
}
\arguments{
\item{n}{An \code{unsigned int} indicating the number of observations to generate.}

\item{mu}{A \code{double} indicating the mean of the normal.}

\item{sigma}{A \code{double} indicating the standard deviation of the normal.}

\item{a}{A \code{double} that is the lower bound of the truncated normal.}

\item{b}{A \code{double} that is the upper bound of the truncated normal.}
}
\description{
Enables sampling from a truncated normal
}
\examples{
rtruncated_normal(10, 1, 1, -2, 2)
}
