% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_local.R
\name{run_squeezemeta}
\alias{run_squeezemeta}
\title{Run SqueezeMeta locally using processx}
\usage{
run_squeezemeta(samples_file, input_dir, project_name, workdir)
}
\arguments{
\item{samples_file}{Path to samples file (-s)}

\item{input_dir}{Directory with input reads (-f)}

\item{project_name}{Project name (-p)}

\item{workdir}{Working directory where project folder will be created}
}
\value{
processx process object
}
\description{
Run SqueezeMeta locally using processx
}
