% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sarma_components}
\alias{sarma_components}
\title{Determine parameter expansion based upon objdesc}
\usage{
sarma_components(objdesc)
}
\arguments{
\item{objdesc}{A \code{vec} with the appropriate sarima object description}
}
\value{
A \code{vec} with the structure:
\describe{
\item{np}{Number of Non-Seasonal AR Terms}
\item{nq}{Number of Non-Seasonal MA Terms}
\item{nsp}{Number of Seasonal AR Terms}
\item{nsq}{Number of Seasonal MA Terms}
\item{ns}{Number of Seasons (e.g. 12 is year)}
\item{p}{Total number of phi terms}
\item{q}{Total number of theta terms}
}
}
\description{
Calculates the necessary vec space needed to pad the vectors
for seasonal terms.
}
\keyword{internal}
