% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sarma_expand}
\alias{sarma_expand}
\title{Expand Parameters for an SARMA object}
\usage{
sarma_expand(params, objdesc)
}
\arguments{
\item{params}{A \code{vec} containing the theta values of the parameters.}

\item{objdesc}{A \code{vec} containing the model term information.}
}
\value{
A \code{field<vec>} of size two as follows:
\itemize{
  \item AR    values
  \item THETA values
}
}
\description{
Creates an expanded PHI and THETA vector for use in other objects.
}
\details{
The \code{objdesc} is assumed to have the structure of:
\itemize{
\item AR(p)
\item MA(q)
\item SAR(P)
\item SMA(Q)
\item Seasons
}
}
\examples{
# p, q, P, Q, 1, s, i, si
m = sarma_expand(c(0.5,.2,0,.1,.92,.83,.42,.33,.12), c(2,2,2,3,1,12,0,0))
}
\keyword{internal}
