% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sarma_params_construct}
\alias{sarma_params_construct}
\title{Efficient way to merge items together}
\usage{
sarma_params_construct(ar, ma, sar, sma)
}
\description{
Efficient way to merge items together
}
\keyword{internal}
