% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{setModelPars}
\alias{setModelPars}
\title{set parameter values for a trajectory model}
\usage{
setModelPars(model, par = NULL)
}
\arguments{
\item{model}{an object of class \code{Model}}

\item{par}{a named parameter vector (or \code{NULL} for default values)}
}
\description{
transformed parameter values are returned
}
