% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotype_sharing.R
\name{share_table}
\alias{share_table}
\title{Computes sharing level of unique clones among samples}
\usage{
share_table(populations_list)
}
\arguments{
\item{clonotype_list}{list of character vectors}
}
\value{
df
}
\description{
Counts the number of times each clonotype appears in a sample vector
}
\examples{

rand_rep <- function(n, alphabet = LETTERS[1:10])
   sample(alphabet, n, replace = TRUE)

rep_list <- replicate(3, rand_rep(rpois(1, 10)), simplify = FALSE)

share_table(rep_list)

}
