% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theoretical_wv}
\alias{theoretical_wv}
\title{Model Process to WV}
\usage{
theoretical_wv(theta, desc, objdesc, tau)
}
\arguments{
\item{theta}{A \code{vec} containing the list of estimated parameters.}

\item{desc}{A \code{vector<string>} containing a list of descriptors.}

\item{objdesc}{A \code{field<vec>} containing a list of object descriptors.}

\item{tau}{A \code{vec} containing the scales e.g. \eqn{2^{\tau}}{2^tau}}
}
\value{
A \code{vec} containing the wavelet variance of the model.
}
\description{
This function computes the summation of all Processes to WV (haar) in a given model
}
\section{Haar Wavelet Derivation Information}{

For more information, please see: \href{http://smac-group.com/computing/process-to-haar-wavelet-variance-formulae}{Supported Haar Wavelet Formulae} (Internet Connection Required).
}

\examples{
model = AR1(.3,2) + RW(.21) + DR(.001)
ntau = 8
tau = 2^(1:ntau)
wv.theo = theoretical_wv(model$theta, model$desc, model$objdesc, tau)
}
\keyword{internal}
