% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/inference.cpp, src/inference.h
\name{theta_ci}
\alias{theta_ci}
\title{Generate the Confidence Interval for Theta Estimates}
\usage{
theta_ci(theta, A, v_hat, omega, alpha)
}
\arguments{
\item{theta}{A \code{vec} containing the estimates}

\item{A}{A \code{mat} that is the first derivative matrix.}

\item{v_hat}{A \code{mat} that is the bootstrapped V matrix}

\item{omega}{A \code{mat} that is the inverse of the diagonal V matrix.}

\item{alpha}{A \code{double} that contains the confidence level.}
}
\value{
A \code{mat} that has the first column
}
\description{
Create an Asymptotic CI for the Theta Estimates.
}
\keyword{internal}
