% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{transformPar,TransformedModel-method}
\alias{transformPar,TransformedModel-method}
\title{transform parameter values}
\usage{
\S4method{transformPar}{TransformedModel}(model, par, inverse = FALSE)
}
\arguments{
\item{par}{a named vector (or list) of parameters to be transformed}

\item{inverse}{logical, if TRUE, map from the TransformedModel parameters to the original model parameters; if FALSE, map from the original model parameters to the TransformedModel parameters}

\item{trans}{a list of Model objects defining parameter transformations}
}
\value{
a named list of transformed parameter values
}
\description{
Map parameter values of a transformed model to/from the parameter values of the original model.
}
