\name{undocumented-functions}
\alias{createColorFunction}
\alias{createColorFunction,ANY-method}
\alias{createMatForLayout}
\alias{createMatForLayout,MrpVisualizer-method}
\alias{createStatemapsFunction}
\alias{createStatemapsFunction,MrpStateVisualizer-method}
\alias{dataGenericAugment}
\alias{dataGenericAugment,mrp-method}
\alias{dataGenericAugment,NWayData-method}
\alias{dataLookup}
\alias{dataLookup,NWayData-method}
\alias{dataLookup,mrp-method}
\alias{dataRescale}
\alias{dataRescale,mrp-method}
\alias{dataRescale,NWayData-method}
\alias{fitMultilevelModel}
\alias{getDataLength}
\alias{getDesignEffect}
\alias{getDesignEffectByCell}
\alias{getN}
\alias{getNEffective}
\alias{getNumberWays,NWayData-method}
\alias{getYbarWeighted}
\alias{hasMultilevelModel}
\alias{newMrpStateVisualizer}
\alias{newMrpVisualizer}
\alias{placeCharts}
\alias{placeCharts,MrpVisualizer-method}
\alias{placeLabels}
\alias{placeLabels,MrpVisualizer-method}
\alias{placeMisc}
\alias{placeMisc,MrpVisualizer-method}
\alias{placeMisc,MrpStateVisualizer-method}
\alias{prepareLayout}
\alias{prepareLayout,MrpVisualizer-method}
\alias{setColorFunction}
\alias{setColorFunction,MrpStateVisualizer-method}
%\alias{setFormula}
\alias{setFormula,mrp-method}
\alias{setPlotFunction}
\alias{setPlotFunction,MrpStateVisualizer-method}
%\alias{setPopulation}
\alias{setPopulation,mrp-method}
\alias{visualize}
\alias{visualize,MrpVisualizer-method}
\title{All undocumented methods}
\description{
All undocumented methods are here so there's a place to track them.
}