% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMWM.R
\name{update.gmwm}
\alias{update.gmwm}
\title{Update (Robust) GMWM object for IMU or SSM}
\usage{
\method{update}{gmwm}(object, model, ...)
}
\arguments{
\item{object}{A \code{gmwm} object.}

\item{model}{A \code{ts.model} object containing one of the allowed models}

\item{...}{Additional parameters (not used)}
}
\value{
A \code{gmwm} object with the structure: 
\describe{
 \item{estimate}{Estimated Parameters Values from the GMWM Procedure}
 \item{init.guess}{Initial Starting Values given to the Optimization Algorithm}
 \item{wv.empir}{The data's empirical wavelet variance}
 \item{ci.low}{Lower Confidence Interval}
 \item{ci.high}{Upper Confidence Interval}
 \item{orgV}{Original V matrix}
 \item{V}{Updated V matrix (if bootstrapped)}
 \item{omega}{The V matrix inversed}
 \item{obj.fun}{Value of the objective function at Estimated Parameter Values}
 \item{theo}{Summed Theoretical Wavelet Variance}
 \item{decomp.theo}{Decomposed Theoretical Wavelet Variance by Process}
 \item{scales}{Scales of the GMWM Object}
 \item{robust}{Indicates if parameter estimation was done under robust or classical}
 \item{eff}{Level of efficiency of robust estimation}
 \item{model.type}{Models being guessed}
 \item{compute.v}{Type of V matrix computation}
 \item{augmented}{Indicates moments have been augmented}
 \item{alpha}{Alpha level used to generate confidence intervals}
 \item{expect.diff}{Mean of the First Difference of the Signal}
 \item{N}{Length of the Signal}
 \item{G}{Number of Guesses Performed}
 \item{H}{Number of Bootstrap replications}
 \item{K}{Number of V matrix bootstraps}
 \item{model}{\code{ts.model} supplied to gmwm}
 \item{model.hat}{A new value of \code{ts.model} object supplied to gmwm}
 \item{starting}{Indicates whether the procedure used the initial guessing approach}
 \item{seed}{Randomization seed used to generate the guessing values}
 \item{freq}{Frequency of data}
}
}
\description{
Provides a way to estimate different models over the previously estimated 
wavelet variance values and covariance matrix.
}
\details{
The motive behind this function is to allow for reuse of the \code{\link{gmwm}}
object's computation of the wavelet variance and covariance matrix. The
function permits this by allowing for the underlying time series model to
be changed at will. As a result, the function is particular useful for 
working with large time series objects. Alternatively, one can also use this
function to supply a custom diagonal weighting matrix by modifying the
\code{\link{gmwm}} object.
}
\examples{
# AR
set.seed(1336)
n = 200
exact.model = AR1(phi=.99, sigma2 = 0.01) + WN(sigma2=1)
data = gen_gts(n, exact.model)

# Create an initial model that is not accurate
bad.model = gmwm(AR1(), data = data)

# Models can contain specific parameters e.g.
updated.model = update(bad.model, exact.model)

# Or...
updated.model.guided = update(bad.model, AR1()+AR1())
}
