% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_drift}
\alias{var_drift}
\title{Variance DR}
\usage{
var_drift(omega, n_ts)
}
\arguments{
\item{omega}{A \code{double} corresponding to variance of drift.}

\item{n_ts}{An \code{int} indicating the length of the time series.}
}
\value{
A \code{vec} containing the variance of the drift.
}
\description{
This function computes the variance of a drift process.
}
\examples{
var_drift(1, 200)
}
\keyword{internal}
