% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_to_set}
\alias{vector_to_set}
\title{Conversion function of Vector to Set}
\usage{
vector_to_set(model_str)
}
\arguments{
\item{x}{A \code{vec<vec<string>>} that contains a list of model descriptors.}
}
\value{
A \code{set<vector<string>>} that contains the list of unique models.
}
\description{
Converts a vector into a set
}
\keyword{internal}
